quantity 자료형 및 SD/GR 상세 수정

+++ src/main/java/com/youthcase/orderflow/pr/dto/PurchaseRequestCreateDto.java
@@ -5,6 +5,6 @@ import java.time.LocalDate;
 
 public record PurchaseRequestCreateDto(
         String gtin,
-        int qty,
+        Long qty,
         LocalDate expectedDate // 선택값이면 null 허용
 ) {}
